requirements.txt
git+https://github.com/Rapptz/discord.py
datetime
openpyxl
ezsheets
discord.py[voice]
youtube-dl